refactor: extract business logic from routers, consolidate duplicates, remove dead code
- Extract AI curation orchestration into services/tournament.py (was duplicated
  between create_tournament and regenerate_tournament endpoints)
- Move pair selection algorithm (135 lines) from routers/movies.py to new
  services/pair_selection.py for testability
- Consolidate duplicated movie upsert logic into shared build_movie_upsert()
  in services/pool.py (was maintained separately in pool.py and expand.py)
- Delete dead models.py re-export shim (migration complete, no importers)
- Remove unused ELO_K_FACTOR/ELO_DEFAULT_RATING config (elo.py uses own constants)
- Remove unused outcome_to_scores() function and its tests
